Pearce to Pearwin.
I hope to make alex.pearwin.com the canonical URL soon, but I need to
spend a bit of time to be sure that redirects from alexpearce.me will
work (cool URIs don't change!).

I'm not sure about changing my GitHub username. It's referenced in a
lot of places (on GitHub and elsewhere) and I don't like the idea of
breaking all those.